requirements.txt
psycopg2_binary==2.9.6
pytest==7.2.1
python-dotenv==1.0.1
PyYAML==6.0.2
coverage
inquirerPy==0.3.4
fastapi
pylint
requests==2.32.3
tabulate
uvicorn